backend/requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
blinker==1.7.0
boto3==1.28.7
botocore==1.31.7
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.0.3
django-allauth==0.52.0
django-apscheduler==0.6.2
django-dotenv==1.4.2
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
django-prometheus==2.3.1
django-redis==5.3.0
django-restframework==0.0.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.3
drf-yasg==1.21.6
ecs-logging==2.0.2
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
exceptiongroup==1.1.2
filebeat==0.0.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
logstash==0.1.dev0
Markdown==3.4.3
MarkupSafe==2.1.3
mysqlclient==2.2.0
numpy==1.25.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
prometheus-client==0.17.1
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-logstash==0.4.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.29.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.8.10
s3transfer==0.6.1
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==3.0.1
wsproto==1.2.0
